First version of GNATBus
This patch also modifies qemu-plugin to export the same interface.

Part of K125-030.

--- merged with ---

First version of the Qemu Plug-in interface

 - Device registration
 - I/O memory registration
 - I/O read/write
 - DMA read/write
 - IRQ
 - Events

An example is provided in directory plugin/sample/

Part of K125-030.

--- merged with ---

Plug-in: Change type of size parameter for DMA operations

--- merged with ---

GNATbus over UNIX domain socket and Named pipe

dev_set_{un}blocking() was in fact not required because the IO thread is
either not started or in suspended state when we call
gnatbus_receice_packet_sync().

Part of LA18-024.