fix compilation error in virtio-9p-local.c on old Linux Kernels
The AT_FDCWD and AT_SYMLINK_NOFOLLOW macros were only introduced
in 2006.  Older kernels do not provide these macros in fcntl.h.
This patch just defines them locally to their actual value if
not defined.  As the comment hints, it's a bit iffy, but ultimately
OK in this case, because the use we are making is such that the
values end up being discarded/unused.