ppc: Cancel reservation on CPU switch
The following test case with 2 CPUs proves that the semantics of the
"lwarx" and "stwcx." instructions are not preserved by the emulation.
The test case does the following :
	- CPU0: perform lwarx on a memory location
	- CPU1: perform lwarx on the same memory location
	- CPU0: perform stwcx. on the location (the operation succeeds
	  while it is supposed to fail since the reservation is supposed
	  to be lost)

This patch fixes this problem in a very simple manner by removing the
reservation of a CPU every time it is scheduled (in cpu_exec()). While
this is a harsh workaround, it does not affect the guest code much
because reservations are usually held for a very short time (an lwarx is
almost always followed by an stwcx. a few instructions below).
Therefore, in most cases, the reservation will be taken and consumed
before a CPU switch occurs, however in the rare case where a CPU switch
does occur between the lwarx and its corresponding stwcx. this patch
solves a potential erroneous behavior of the synchronization
instructions.

Signed-off-by: Elie Richa <richa@adacore.com>