Little hack to improve socket latency
On Windows we have two event functions, select for sockets and
WaitForMultiplesObjects (WFMO) for the rest. WFMO is blocking while
select is not, therefore when a socket receive a packet we have to wait
for WFMO to reach the end of timeout. This adds latency in socket
communication.

By creating a Windows event handler attached to the socket, WFMO will
return as soon as there something to read on a socket.

We still have to watch to see if this brings any drawbacks.

Part of KC09-008.