Add vx653 boot loader (prototype)
Each vx653 build produces multiple binaries (ELF) and a boot.txt file.
In boot.txt there is a list of binaries that should be loaded in memory.

This boot loader is not working, the process in more complex that just
loading each ELF binary and involves data structure for memory mapping
configuration (see bootTxtFileLoad and bootElfModuleLoad in vx654).

I leave it there as a prototype, we may be able to improve it in the
future.

Part of L926-022.