Add HostFS device
This device implements a bridge for file syscalls: open, read, write, close.

Part of L613-027.

--- merged with ---

Hostfs: add shutdown syscall

--- merged with ---

hostfs: create the 64bits device

This creates a 64bits device when the target is in 64bits mode.

So this doesn't change the device with 32bits targets but double
the register size when using a 64bits target (eg: aarch64).

TN: Q323-013

2020-02-11  KONRAD Frederic  <konrad@adacore.com>
      * hw/adacore/hostfs.c: fix against include refactoring (SB14-048).
2019-02-01  KONRAD Frederic  <konrad@adacore.com>
      * hw/adacore/hostfs*: convert to realize.
2018-07-20  KONRAD Frederic  <konrad@adacore.com>
      * moved the sources and headers from hw/ to hw/adacore because of the
        broken traces.
      * added the cause argument to qemu_system_reset_request as required.
      * replaced off_t by int for portability.
      * moved hostfs compilation from common-obj-y to obj-y as it becomes
        target dependent because of the bitness (32bits vs 64bits).

Change-Id: Ib677e53545bf61f50dc8b3c064292067bad093b0