ARM semihosting: print console output on STDOUT instead of STDERR
Required if we use semihosting as default output for ARM BB targets.

Part of P808-009.

--- merged with ---

ARM Semihosting: Add -semihosting-console option

to be able to define a chardev for the semihosting (file, tcp, etc.).

If the option is not used, the semihosting console will print in the
first serial chardev, if any, or STDERR.