ppc64: introduce virt platform
This introduces the virt platform.
It is described by the qsp.dtb and is enough to boot the vxWorks kernel.

At the init, it patches the kernel with the memory size passed by the -m
parameters the same way u-boot does it.

If no device tree has been provided it will automagically create one and
provide it to the OS.

2019-02-28  KONRAD Frederic  <konrad@adacore.com>
        * hw/ppc/vxworks-virt.c: added load_elf argument.

TN: Q831-018
no-precommit-check
Change-Id: I1f2e4248feaf1a6e97f3d0c5294f4cadfbd39a00