introduce vxworks-virt platform
This introduces the vxworks-virt platform.
It is described by the qsp.dtb and is enough to boot the vxWorks kernel.

At the init, it patches the kernel with the memory size passed by the -m
parameters the same way u-boot does it.

If no device tree has been provided it will automagically create one and
provide it to the OS.

TN: Q831-018
no-precommit-check

Change-Id: I1f2e4248feaf1a6e97f3d0c5294f4cadfbd39a00