qemu-char: make qemu_chr_write_log static
This makes qemu_chr_write_log function static.
The static attribute was wrongly dropped during a rebase so devices can use
it. But finally it was a bad idea.

This fixes also a warning because qemu_chr_write_log was not defined.

NOTE: must be merged with 0e2fe33976127fc91493c679a9dcf7723da942a2

no-precommit-check

Change-Id: I928f2109d150ed53b7394b8339e7042bd7c8e5dd