ppc: fix conditional branch with bo = 20
The GNATCov testsuite has been broken during the rebase. It seems that:
`bcl 20, xx, xx` instruction is not working anymore.

When type == BCOND_IM no jump was done which leads to this bug.
So this just fix this wrong behavior.

NOTE: A testcase for conditional branch will be added to GNATemu to keep an
      eye on that.
NOTE2: to be merged with the commit which introduced that bug.

TN: Q905-029
no-precommit-check

Change-Id: I80f4c5b47cf522eb25323ed17a32b72ac5707271