win64: give a chance to vcpu thread to run workaround
This is a workaround for the internal issue we have on windows 64 hosts.

The bug is reproducable on windows64 hosts by starting qemu with one or more
tcp sockets. It results in a huge slow down compaired to the same command line
without any tcp socket.

Technically the iothread become busy polling TCP's fds because g_poll(..)
returns immediately whatever the timeout is and even if there are no data on
the socket.

The results is that the VCPU thread has no chances to start because it's
kicked before it has any chance to enter the execution loop.

So this patch adds a little delay of 1ms if the main-loop spins more than 1000
time. NOTE that we add time to the timeout on linux in the same case but as
timeout as no effect on TCP's fds we have to sleep.

Fortunately a cleaner fix is released in the next qemu revision (2.9.0) but
digging it out is not a good idea as we are not sure it is stable with all the
other patches. So this is a temporary workaround.

NOTE: We can drop that as soon as we reach 2.9.0 version

TN: Q921-018
no-precommit-check

Change-Id: I70ac7b2c366eaaf156a5f1ee620610dfaec1b066