gnatbus: fix write memory leak
In GNATBus the write packet has a uint8_t data[] as a last element which means
that we need to dynamically allocate the packet if we doesn't want to spend
time to send extra data. But this packet was never desallocated which caused a
memory leak. So just fix that.

A better fix would be to fix the packet to have a (void *) and just pass the
data as a pointer. So we save one copy + one malloc per transaction which is
great. But keep this change for later after 18.0.

TN: Q907-020
no-precommit-check

Change-Id: If63ddd595752083da9a959ae5945f93c4f794c45