win32: exit deadlock workaround
There is a deadlock happening rarely in KERNEL32!ExitProcess.

According to MSDN:
If one of the terminated threads in the process holds a lock and the
DLL detach code in one of the loaded DLLs attempts to acquire the same
lock, then calling ExitProcess results in a deadlock. In contrast, if
a process terminates by calling TerminateProcess, the DLLs that the
process is attached to are not notified of the process termination.

So this patch works around the deadlock by calling TerminateProcess
instead of simply exiting the main(..) function.

2021-08-05  Frederic Konrad  <konrad@adacore.com>
        * moved from vl.c to main.c

TN: QA10-046
no-precommit-check
Change-Id: Ia4e641cf0f1a4daf29994a2dacc92f9d68388625

Change-Id: I827090c2c1640b926caf2af01a029044ca6d6f39