i386: mingw: powl, logl workaround
This is a workaround for x86 target under MinGW. The powl and the logl
function doesn't behaving well when the FPU has the wrong precision.

Reinitializing the FPU Control register fix this issue.

--- merged with ---

target-i386: mingw: powl, logl workaround 2nd round

Some compiler optimisation was using the fpu between the assignation
and the actual utilisation of the variable which leads to worse result
than before.

Let's fix that by moving the "finit" workaround before the assignation
and make everybody volatile to avoid any issues.

TN: QB30-019
Change-Id: I8d16e6237d4341f7e3e2645393e8d0f308c19947