riscv: add a dummy xml file
Reading core register is not allowed when we don't have the target description
xml file.. So lets fix that by adding an XML file. The way GDB works today is
obscur it will trigger a warning when we try to specify any register in the
xml file:

warning: Target-supplied registers are not supported by the current
         architecture

It's not an issue as they seems hardcoded inside GDB..

With that we are able to reply GDB when it tries to get the MISA register
before doing a call command.

TN: R703-048
Change-Id: Icba0a868142f348f6a96236f25d87ed2549526da