leon3: make the gptimer irq base configurable
The gptimer irq base must be configurable as there are several variant of
those SOC with different IRQ location. The machine option is not a good
idea in our case.

The best idea here is that the board have a device which has the option
which need to be put in place..

Adding this to the command line will set the irq base to X:
  --global driver=leon3-conf,property=gptimer-irq-base,value=X

TN: S102-004