win32: increase the main thread priority
When working on heavy loaded machine with an OS doing live idle loop such as:
  loop:
    b loop

And waiting for a timer to fire the main-loop thread have less chance to
execute so the timer interruptions are delayed. So lets increase the main
thread priority.

A major drawback of this is that doing too much IO such as network packet
might slow down the execution of the guest code.

TN: S702-019
Change-Id: I756e49bc40eed868c2474cb0a2aa3eea729544c6