net/phy/marvell.c: keep the link state during the soft reset
When a soft reset happen the phy resets all the registers. The network
backend doesn't know about that and won't update the link. So in the end the
guest driver will think that the link is down. Keep the LINK bit from the
STATUS register and update the link status with that after having reset the
registers.

to be squashed with:
  7bb5e7a4e92f24ef2172fed171a7eccbb8e48df0

TN: S731-009
Change-Id: Ibe02260ae9cb3b8ecc5b72d7aeed9ec450bcbb09