mingw32: reinitialize the fpu when enterring the tcg thread.
This is a workaround for MinGW where some maths functions are not working
correctly when called from a newly created thread.  The precision of the FPU
is bad when the thread is created with begin_thread_ex.  This ends up with a
precision loss within the guest floating point instruction.

TN: SA28-077
Change-Id: Ia1c366a92fa642da60c31f56a1fac2d43f063816