m68k: implements the timer of the mc68901 device
This implements the four timers of the mc68901 device.  The four timers have
a specific GPIO line (qemu_irq) to trigger a side effect on anything which
is bound to it when the decrementer is passing from 1 to 0. The support is
not complete here are the limitations:
  - The interrupts are not modeled.
  - The input signals for TimerA and TimerB are missing so the PWM mode is
    not working.  Using it is triggering a GUEST_UNIMP trace.

The rest should work correctly.

TN: T108-046, SA29-066
Change-Id: I2e5eba817509dfb9f33a38075110f9f53b116568