SiFive U boards: swap the CPU cluster ids
Right now only the e_cluster is automatically attached to GDB remote.
Debugging the CPUs of the u_cluster requires a few extra GDB commands.

Setting id 0 to the u_cluster means it will be attached first in GDB
remote and therefore available in right away for debug. Since the
run-time only uses cluster_u, this is the best option for us.

2021-07-15  Fred Konrad  <konrad@adacore.com>
        * Replaced the polarfiresoc.c BSP with the upstream BSP.

Part of T330-017 and T603-016

Change-Id: Ie8b47bb08de1f40a672d3aebdeb8f26430827312