ppc: restore the MSR correctly during rfi
PowerPC programmer's reference guide state:

For 32-bit implementations, when an exception occurs, bits 1-4 and 10-15
of SRR1 are loaded with exception-specific information and bits 0, 5-9,
and 16-31 of MSR are placed into the corresponding bit positions of SRR1.

So makes sure that the exception-specific informations are dropped when
restoring MSR (masking the 0x783F0000 field).

TN: UB19-038

rebase on 7.0.0
TN: QB08-037

Change-Id: I1444d06c967cfcae9cd13b51e4856fdac9735276