Makefile: avoid updating meson-buildoptions.sh by default
On Windows, make will always try to update meson-buildoptions.sh
when running the default target. This is probably due to timestamps
being wrong in some cases. However, this will triggered configure again
which is always failing (mostly because of files already existing).
As we don't want to update this script anyway, simply disable this part.

TN: QB08-037
Change-Id: I9d241dc218d16979603cf4d25680553990d7a794