io: switch completely to Windows API for socket watch on win32
The current implementation is using a mixed of select and the Windows
API (WSASelectEvent, etc).
As select doesn't clear the Windows events, WSAENumNetworkEvents must be
called before select to do it.
However, as this operation isn't made atomically, some events might be
reset before being retrieved by select. Resulting in them being skipped
by the polling.

In order to avoid this issue, this patch improves socket watch to be
based solely on Windows API.

Note that because several GSources might be created for the same HANDLE,
this implementation has no way to know on which GSources an event must be
returned. However, it seems that only two GSources are created, a master
one with just "G_IO_HUP" and the child one which aims to retrieve all
the events (ie with "G_IO_IN", "G_IO_OUT", etc). So, to avoid the master
GSource receiving events meant for the child, we aren't calling
WSAEnumNetworkEvents for it (or for any GSource having only G_IO_HUP as
condition). As G_IO_HUP was never returned by the previous, I supposed
that it should be safe to do so.

TN: V812-020
Change-Id: I037fbe93e0641894b0096c39ace5d0d68db8e2e3