meson: disable -static with enable_static
Configuring with --static allows us to link all the Qemu dependencies
statically. However, by default it adds a LDFLAGS -static which will
also statically linked libc, libm and other system libraries which we
don't want.

Configure is also testing -static-pie flag when both --static and
--enable-pie are provided. However, this flag is not supported in
all our distributions (at least ubuntu18.04 and rhES7). As this
test is no longer needed for us, remove it.

This is a workaround. The right patch would probably to introduce a
--partial-static options or find the right LDFLAGS.

FYI, here is the dynamic libs with qemu-7:
 0x0000000000000001 (NEEDED)             Shared library: [libfdt.so.1]
 0x0000000000000001 (NEEDED)             Shared library: [librt.so.1]
 0x0000000000000001 (NEEDED)             Shared library: [libutil.so.1]
 0x0000000000000001 (NEEDED)             Shared library: [libm.so.6]
 0x0000000000000001 (NEEDED)             Shared library: [libdl.so.2]
 0x0000000000000001 (NEEDED)             Shared library: [libpthread.so.0]
 0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]

And here is the dynamic libs with qemu-8 with this option:
 0x0000000000000001 (NEEDED)             Shared library: [libm.so.6]
 0x0000000000000001 (NEEDED)             Shared library: [libutil.so.1]
 0x0000000000000001 (NEEDED)             Shared library: [libpthread.so.0]
 0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]

TN: T317-002