qemu-io-cmds: implement clock_gettime on Windows
clock_gettime is not available on our MinGW setup. We used to revert the
commit introducing it, but this is no longer possible: other commits
have added some clock_gettime calls as well.
Instead, introduce a quick implementation of clock_gettime for Windows
based on gettimeofday. This is less precise and might not always work.
But we don't care about benchmarking at the moment.

TN: T514-018