qemu-traces: add traces for arm targets
The main changes is to pass TB_EXIT_NOPATCH only if the traces are active
to avoid a performance penalty (~7%).

Starting in qemu-9, ARM havs moved to relative PCs in translation blocks.
According to the doc in include/exec/exec-all.h, this means that
the PC cannot be retrieved solely with the TB anymore. It has to be
taken from the ENV variable.

However, some parts of our trace algorithms (exec_trace_after_exec)
requires to have a fully available pc in translation blocks. Thus,
we can't enable this new feature for ARM.

TN: O520-026, U730-008, U901-009