qemu-traces: disable relative PC features when traces are enabled
ARM and x86 cpus have moved to relative PCs in translation blocks.
According to the doc in include/exec/exec-all.h, this means that
the PC cannot be retrieved solely with the TB anymore. It has to be
taken from the ENV variable.

However, some parts of our trace algorithms (exec_trace_after_exec)
requires to have a fully available pc in translation blocks. Thus,
we can't enable this new feature for the architecture supporting traces
(= arm).