configure: remove -static-pie test
Configure is testing -static-pie flag because, when --static and
--enable-pie are both provided, meson.build will try to add -static-pie
to LDFLAGS. However, this flag has been removed in our builds (as we
don't want to statically link the system libraries: see 57b99d9d). Thus,
this test is no longer needed for us.

As some distributions (at least ubuntu18.04 and rhES7) don't seem to
support it, we need to remove it.