meson: force shared libraries for libm and libutil
We prefer using static libraries in our builds to embedded all
dependencies in the final executable. However, system libraries must be
exempted from this rule.

For an unknwon reason, meson refuses to fetch the shared libutil if not
asked for. `static: false` must be forced. Libm being also a system
library to the same just in case.

qemu#31