hw/ppc: correct p3041ds core frequency
This board is exclusively called by VxWorks targets on our
infrastructure. Both vx6 and vx7r2 document states that the board should
be configured with a 1.5Ghz core frequency.

Vx7r2 doesn't seem to care too much about it but vx6 does and will
freeze if the frequency is wrong.

qemu#34