qemu-traces: create internal structure for trace entries
The format of the traces is fixed and must never be changed because of
gnatcov compatibility. But this format doesn't need to restrict how the
entries are stored within QEmu.

Therefore, a new internal structure has been made to record traces;
those traces being converted to the "external" format before being
flushed to the file.

Helper functions are being added for later use. There are encapsulated
within #ifdef for now, to avoid unused warnings.

eng/shared/release#1025
gnatemulator#37