qemu-traces: allow emulation independent traces generation
AARCH64 is now the emulation for both arm and aarch64 targets and thus
whether the target is 32bit or 64bti must be detect by the trace
generator instead of hardcoded.

For now, all targets but ARM ones are still set is_32bit using ifdef.
For arm, we retrieve the first cpu and detect its architecture.

eng/shared/release#1025
gnatemulator#37